+++ triplea/service/graph/analysis/__init__.py
@@ -3,6 +3,7 @@ from triplea.service.graph.analysis.ganalysis import (
     sorted_degree_centrality,
     show_degree_distribution,
     sorted_average_neighbor_degree,
+    get_avg_shortest_path_length_per_node,
 )
 
 from triplea.service.graph.analysis.diameter import graph_diameter
@@ -16,4 +17,5 @@ __all__ = [
     "sorted_average_neighbor_degree",
     "average_neighbor_number",
     "graph_diameter",
+    "get_avg_shortest_path_length_per_node",
 ]